+++ addons/TBMod_Eden/ReadMe.md
@@ -1,0 +1,3 @@
+# TBMod Eden
+
+Developer:     [Eric Ruhland](https://github.com/Er1807)
