+++ addons/TBMod_Eden/config.cpp
@@ -1,0 +1,24 @@
+/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgPatches
+{
+    class TBMod_main
+    {
+        name = "TBMod Eden";
+        
+        units[] = {
+            
+        };
+
+        requiredAddons[] = {"TBMod_main"};
+        addonRootClass = "TBMod_main";
+    };
+};
+
+// Configs
+#include "configs\CfgFunctions.hpp"
+#include "configs\CfgVehicles.hpp"
+#include "configs\Cfg3DEN.hpp"
+#include "configs\ExtendedEventHandlers.hpp"
