+++ addons/TBMod_Eden/configs/CfgFunctions.hpp
@@ -1,0 +1,21 @@
+﻿/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgFunctions
+{
+    class TBMod_main
+    {
+        tag = "TB";
+
+        class functions
+        {
+            file = "\TBMod_Eden\functions";
+            class schauspieler {};
+            class sitAI {};
+            class animationAI {};
+            class keysForAI {};
+            class inkognitoSystem {};
+        };
+    };
+};
