+++ addons/TBMod_Eden/functions/fn_inkognitoSystem.sqf
