+++ addons/TBMod_Eden/functions/fn_sitAI.sqf
