+++ addons/TBMod_ZeusModuls/configs/CfgVehicles.hpp
@@ -1,4 +1,7 @@
-class CBA_Extended_EventHandlers;
+/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
 
 class CfgVehicles
 {
