+++ addons/TBMod_ZeusModuls/configs/ExtendedEventHandlers.hpp
@@ -1,0 +1,14 @@
+/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+
+class Extended_PreInit_EventHandlers
+{
+    class TBMod_Zeusmoduls
+    {
+        init = "call compile preprocessFileLineNumbers '\TBMod_Zeusmoduls\XEH_preInit.sqf'";
+    };
+};
+
+
