+++ addons/TBMod_main/XEH_postClientInit.sqf
@@ -31,7 +31,6 @@ enableSaving [false, false];
     [] spawn TB_fnc_shortcuts;
     [] call TB_fnc_chatCommands;
     [] call TB_fnc_extremMedicMode;
-    [] call TB_fnc_schauspieler;
     [] spawn TB_fnc_miscStuff;
     
     TB_init_done = true;
